requirements.txt
Django>=1.8,<2.0
djangorestframework
django-cors-headers
django-oauth-toolkit
psycopg2
dj-database-url==0.4.2
gunicorn==19.7.1
whitenoise==3.3.1